feat: increase select2 heights as max as possible
